ophalen Playlistdata uit DB gefixt

+++ src/main/java/nl/nusayba/oose/domain/services/PlaylistService.java
@@ -1,35 +1,77 @@
 package nl.nusayba.oose.domain.services;
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
 
 import nl.nusayba.oose.domain.dto.PlaylistDTO;
+import nl.nusayba.oose.domain.dto.PlaylistsDTO;
 import nl.nusayba.oose.domain.interfaces.IPlaylistDAO;
+import nl.nusayba.oose.domain.interfaces.ILoginDAO;
 
-import java.util.List;
+import jakarta.inject.Inject;
+import jakarta.enterprise.context.RequestScoped;
 
-@ApplicationScoped
+@RequestScoped
 public class PlaylistService {
 
-    @Inject
     private IPlaylistDAO playlistDAO;
+    private ILoginDAO loginDAO;
 
-    public List<PlaylistDTO> getAllPlaylists() {
-        return playlistDAO.getAllPlaylists();
+    @Inject
+    public void setPlaylistDAO(IPlaylistDAO playlistDAO) {
+        this.playlistDAO = playlistDAO;
     }
 
+    @Inject
+    public void setLoginDAO(ILoginDAO loginDAO) {
+        this.loginDAO = loginDAO;
+    }
     public PlaylistDTO getPlaylistById(int id) {
         return playlistDAO.getPlaylistById(id);
     }
-
-    public void addPlaylist(PlaylistDTO playlist) {
-        playlistDAO.insertPlaylist(playlist);
+    public PlaylistsDTO getAllPlaylists(String token) {
+        if (isValidToken(token)) {
+            String username = loginDAO.getUserByToken(token).getUser();
+            return playlistDAO.getPlaylist(username);
+        }
+        throw new RuntimeException("Invalid token");
     }
 
-    public void updatePlaylist(PlaylistDTO playlist) {
-        playlistDAO.updatePlaylist(playlist);
+    public PlaylistsDTO addPlaylist(String token, PlaylistDTO playlistDTO) {
+        if (isValidToken(token)) {
+            String username = loginDAO.getUserByToken(token).getUser();
+            playlistDAO.addPlaylist(username, playlistDTO);
+            return playlistDAO.getPlaylist(username);
+        }
+        throw new RuntimeException("Invalid token");
     }
 
-    public void deletePlaylist(int id) {
-        playlistDAO.deletePlaylist(id);
+    public PlaylistsDTO deletePlaylist(String token, int id) {
+        if (isValidToken(token)) {
+            String username = loginDAO.getUserByToken(token).getUser();
+            PlaylistDTO playlist = playlistDAO.getPlaylistById(id);
+            if (playlist.isOwner() && username.equals(playlist.getName())) {
+                playlistDAO.deletePlaylist(id);
+                return playlistDAO.getPlaylist(username);
+            } else {
+                throw new RuntimeException("You do not have permission to delete this playlist");
+            }
+        }
+        throw new RuntimeException("Invalid token");
+    }
+
+    public PlaylistsDTO updatePlaylist(String token, int id, PlaylistDTO playlistDTO) {
+        if (isValidToken(token)) {
+            String username = loginDAO.getUserByToken(token).getUser();
+            PlaylistDTO existingPlaylist = playlistDAO.getPlaylistById(id);
+            if (existingPlaylist.isOwner() && username.equals(existingPlaylist.getName())) {
+                playlistDAO.updatePlaylist(id, playlistDTO);
+                return playlistDAO.getPlaylist(username);
+            } else {
+                throw new RuntimeException("You do not have permission to update this playlist");
+            }
+        }
+        throw new RuntimeException("Invalid token");
+    }
+
+    private boolean isValidToken(String token) {
+        return loginDAO.getUserByToken(token) != null;
     }
 }
